refactor: BaseActivity 구현
- Activity 중복 코드 제거
