+++ app/src/main/java/woowacourse/movie/view/movies/MoviesActivity.kt
@@ -3,20 +3,13 @@ package woowacourse.movie.view.movies
 import android.content.Intent
 import android.os.Bundle
 import android.widget.ListView
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import woowacourse.movie.R
 import woowacourse.movie.model.Movie
+import woowacourse.movie.view.base.BaseActivity
 import woowacourse.movie.view.reservation.result.ReservationResultActivity
 
-class MoviesActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_movies)
-        setWindowInsets()
+class MoviesActivity : BaseActivity() {
+    override fun initView() {
         setMovieListView()
     }
 
@@ -49,13 +42,5 @@ class MoviesActivity : AppCompatActivity() {
                 },
             )
         lvMovie.adapter = movieListAdapter
-    }
-
-    private fun setWindowInsets() {
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
     }
 }
